Added Test cases
