+++ Vitals/BreachFlag.cs
@@ -7,7 +7,7 @@ namespace Vitals
     {
         public static void BreachLogger(string vital,string level,float value)
         {
-            Console.WriteLine(vital + " is " + level + vital + " = " + value);
+            Console.WriteLine(vital + " is " + level +"; "+ vital + " = " + value + "\n\n");
         }
     }
 }
