+++ run.bat
@@ -1,5 +1,5 @@
 pushd "%~dp0"
 
-checker.exe
+Test.exe
 
 popd
